--- a/charts/user/templates/clusterroles.yaml
+++ b/charts/user/templates/clusterroles.yaml
@@ -7,6 +7,7 @@ rules:
     - ""
     resources:
     - pods
+    - configmaps
     - services
     - deployments
     - nodes
@@ -22,6 +23,7 @@ rules:
     - customresourcedefinitions
     verbs:
     - list
+    - get
 {{- with .Values.clusterRoles }}
 {{- toYaml . | nindent 2 }}
 {{- end }}