--- a/pkg/apis/config/arenaconfig.go
+++ b/pkg/apis/config/arenaconfig.go
@@ -70,11 +70,14 @@ func newArenaConfiger(args types.ArenaClientArgs) (*ArenaConfiger, error) {
 	if err != nil {
 		return nil, err
 	}
-	crdNames, err := getClusterInstalledCRDs(apiExtensionClientSet)
-	if err != nil {
-		return nil, err
-	}
+	/*
+		crdNames, err := getClusterInstalledCRDs(apiExtensionClientSet)
+		if err != nil {
+			return nil, err
+		}
+	*/
 	namespace := updateNamespace(args.Namespace, arenaConfigs, clientConfig)
+	log.Debugf("current namespace is %v", namespace)
 	return &ArenaConfiger{
 		restConfig:            restConfig,
 		clientConfig:          clientConfig,
@@ -84,7 +87,7 @@ func newArenaConfiger(args types.ArenaClientArgs) (*ArenaConfiger, error) {
 		arenaNamespace:        args.ArenaNamespace,
 		configs:               arenaConfigs,
 		isDaemonMode:          args.IsDaemonMode,
-		clusterInstalledCRDs:  crdNames,
+		clusterInstalledCRDs:  []string{},
 	}, nil
 }
 
@@ -183,6 +186,7 @@ func updateNamespace(namespace string, arenaConfigs map[string]string, clientCon
 
 func getClusterInstalledCRDs(client *extclientset.Clientset) ([]string, error) {
 	selectorListOpts := metav1.ListOptions{}
+
 	list, err := client.ApiextensionsV1().CustomResourceDefinitions().List(selectorListOpts)
 	if err != nil {
 		return nil, err