1.reduce the execution time of 'arena get' from 2s to 600ms
2.reduce the execution time of 'arena list' from 3s to 700ms
3.reduce the execution time of 'arena top node' from 400ms